+++ package.json
@@ -13,7 +13,6 @@
     "coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@vue/eslint-config-typescript": "11.0.2",
     "axios": "0.27.2",
     "firebase": "9.13.0",
     "pinia": "2.0.23",
@@ -31,12 +30,13 @@
     "@vitejs/plugin-vue": "3.1.2",
     "@vitest/coverage-c8": "0.24.3",
     "@vitest/ui": "^0.24.5",
+    "@vue/eslint-config-typescript": "^11.0.2",
     "autoprefixer": "^10.4.13",
     "c8": "7.12.0",
-    "eslint": "8.27.0",
-    "eslint-config-airbnb-base": "15.0.0",
+    "eslint": "^8.30.0",
+    "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-tailwindcss": "^3.7.1",
     "eslint-plugin-vue": "9.6.0",
