+++ src/components/overview/desktop/index.vue
@@ -1,16 +1,16 @@
 <script lang="ts" setup>
-import NavBar from '@/components/overview/desktop/NavBar.vue';
+import AppBar from '@/components/overview/desktop/AppBar.vue';
 import SideBar from '@/components/overview/desktop/SideBar.vue';
 import Content from '@/components/overview/desktop/Content.vue';
 </script>
 
 <template>
-  <SideBar />
+  <side-bar />
 
   <div class="flex flex-col">
-    <NavBar />
+    <app-bar />
 
-    <Content />
+    <content />
   </div>
 </template>
